Infrastructure and Endpoint

+++ Core/Domain/src/main/java/ir/otragh/core/domain/homes/entities/Home.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 import java.util.LinkedList;
 
 @Getter
-public final class Home extends BaseEntity<Integer> {
+public final class Home extends BaseEntity<Long> {
     private Title title;
     private Description description;
     private Address address;
@@ -22,7 +22,7 @@ public final class Home extends BaseEntity<Integer> {
     private LinkedList<Long> amenities;
     private LocalDateTime lastReservedOnUTC;
 
-    private Home(int id, Title title, Description description, Address address, Money price, LinkedList<Long> amenities) {
+    private Home(long id, Title title, Description description, Address address, Money price, LinkedList<Long> amenities) {
         super(id);
         this.title = title;
         this.description = description;
